ProTo: Moved Proto development tools into separate modules. Other related refactoring.
