+++ lib/proto/tools/ExploreToolsTool.xml
@@ -1,4 +1,4 @@
 <tool id="proto_explore_tools_tool" name="ProTo tool explorer" version="1.0.0"
-  tool_type="proto_generic" proto_tool_module="proto.tools.ManageTools" proto_tool_class="ExploreToolsTool">
+  tool_type="proto_generic" proto_tool_module="proto.tools.ExploreToolsTool" proto_tool_class="ExploreToolsTool">
   <description></description>
 </tool>
