+++ lib/proto/tools/GenerateToolsTool.xml
@@ -1,4 +1,4 @@
 <tool id="proto_generate_tools_tool" name="ProTo tool generator" version="1.0.0"
-  tool_type="proto_generic" proto_tool_module="proto.tools.ManageTools" proto_tool_class="GenerateToolsTool">
+  tool_type="proto_generic" proto_tool_module="proto.tools.GenerateToolsTool" proto_tool_class="GenerateToolsTool">
   <description></description>
 </tool>
